// src/app/api/model/answer-type.ts
/**
 * My App
 * This is an awesome app!
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type AnswerType = 'None' | 'FreeTextAndJustification' | 'FreeText';

export const AnswerType = {
    None: 'None' as AnswerType,
    FreeTextAndJustification: 'FreeTextAndJustification' as AnswerType,
    FreeText: 'FreeText' as AnswerType
};