// src/app/api/model/question.ts
/**
 * My App
 * This is an awesome app!
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AnswerType } from './answer-type';
import { ConferenceInstance3 } from './conference-instance3';
import { QuestionGroup } from './question-group';


export interface Question { 
    id?: string | null;
    questionGroups: Array<QuestionGroup>;
    conference: Array<ConferenceInstance3>;
    question: string;
    answerType: AnswerType;
}
export namespace Question {
}